Reset state on logout
Fixes #206

Although ugly, reloading the window considerably reduces the chance of data leak and also resolves any memory leaks.
