+++ app/renderer/views/Login.js
@@ -1,11 +1,9 @@
-import {remote} from 'electron';
-import {centerWindow} from 'electron-util';
 import React from 'react';
 import {Subscribe} from 'unstated';
 import Progress from 'components/Progress';
 import LoginView from 'components/LoginView';
 import loginContainer from 'containers/Login';
-import {loginWindowSize} from '../../constants';
+import {setLoginWindowBounds} from '../util';
 import NewPortfolio from './NewPortfolio';
 import LoginBox from './LoginBox';
 import CreatePortfolio from './CreatePortfolio';
@@ -13,21 +11,6 @@ import RestorePortfolio from './RestorePortfolio';
 import ForgotPassword from './ForgotPassword';
 import LoggingIn from './LoggingIn';
 import './Login.scss';
-
-const setLoginWindowBounds = () => {
-	const win = remote.getCurrentWindow();
-	win.setFullScreen(false);
-	win.setFullScreenable(false);
-	win.setResizable(false);
-	win.setMaximizable(false);
-	win.setMinimumSize(loginWindowSize.width, loginWindowSize.height);
-	centerWindow({
-		size: {
-			width: loginWindowSize.width,
-			height: loginWindowSize.height,
-		},
-	});
-};
 
 class Login extends React.Component {
 	componentDidMount() {
@@ -58,7 +41,7 @@ class Login extends React.Component {
 							<div className="window-draggable-area"/>
 							<Progress className="login-progress" value={login.state.progress}/>
 							<div className="is-centered">
-								<img className="hyperdex-icon" src="/assets/hyperdex-logo-text.svg" width="130"/>
+								<img className="hyperdex-icon" src="/assets/hyperdex-logo-text.svg" width="130" height="113"/>
 								<LoginViews/>
 							</div>
 						</div>
